app/requirements.txt
ecs-logging==2.0.2
dynaconf==3.2.0
fastapi==0.115.6
gdown==5.2.0
grpcio==1.67.1
mysql-connector-python==9.1.0
numpy==1.26.3
pandas==2.2.3
pillow==10.2.0
pymilvus==2.5.0
python-dotenv==1.0.1
python-multipart==0.0.19
requests==2.32.3
torch==2.5.1
torchvision==0.20.1
tqdm==4.67.1
uvicorn==0.32.1
pydantic_settings==2.7.0
facenet-pytorch